izmjena baze podataka

+++ database/migrations/2025_07_09_104545_create_orders_table.php
@@ -11,29 +11,31 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->string('OrderNumber');
-            $table->date('OrderDate');
-            $table->text('Description');
-            $table->string('Status');
-            $table->string('CurrentEmployee');
-            $table->string('BojaDuzinaProvodnika');
-            $table->string('Pakovanje');
-            $table->string('AtestPaketa');
-            $table->string('CeOznaka');
-            $table->string('KlasaOpasnosti');
-            $table->string('UNBroj');
-            $table->string('RokIsporuke');
-            $table->string('DatumPredaje');
-            $table->string('DatumPrijema');
-            $table->string('Napomena');
+        if (!Schema::hasTable('orders')) {
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->string('OrderNumber');
+                $table->date('OrderDate');
+                $table->text('Description');
+                $table->string('Status');
+                $table->string('CurrentEmployee');
+                $table->string('BojaDuzinaProvodnika');
+                $table->string('Pakovanje');
+                $table->string('AtestPaketa');
+                $table->string('CeOznaka');
+                $table->string('KlasaOpasnosti');
+                $table->string('UNBroj');
+                $table->string('RokIsporuke');
+                $table->string('DatumPredaje');
+                $table->string('DatumPrijema');
+                $table->string('Napomena');
 
-            $table->unsignedBigInteger('ProizvodId');
-            $table->foreign('ProizvodId')->references('id')->on('products')->onDelete('cascade');
+                $table->unsignedBigInteger('ProizvodId');
+                $table->foreign('ProizvodId')->references('id')->on('products')->onDelete('cascade');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
